[py2py3] stdlib - REST_t matches new urllib interface
Before:
* test REST api with 'list?int=a' and urllib2.
  - HTTPError exception catched by self.assertRaises.
  - No introspection
* test REST api with 'list1?int=a' and urllib
  - no exception raised
  - introspection of the response with self.assertEqual

Futurize: replace py2 version of urllib and urllib2 with
python3 urllib library, which behaves as urllib2 in py2,
which means that in case of an http error, a HTTPError
exception is raised.

This means that we need to manually change the code that
takes care of checking if a request returns an error.

We decided to keep both 'list?int=a' and 'list1?int=a',
which raise the same exception, but provide a different messages.

We keep both self.assertRaises and exception introspection.